Refactory de Register component
